% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean/create \code{DATE} variable}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{a dataframe from the NOAA Significant Earthquakes dataset}
}
\value{
A dataframe with new \code{DATE} variable
}
\description{
Create a DATE variable in a NOAA Significant Earthquakes dataset dataframe.
}
