% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean \code{LOCATION_NAME} variable}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{a dataframe from the NOAA Significant Earthquakes dataset}
}
\value{
A dataframe with cleaned version of \code{LOCATION_NAME} variable
}
\description{
Cleans the \code{LOCATION_NAME} variable in a NOAA Significant Earthquakes dataset dataframe.
}
