% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-timeline-label.R, R/geom-timeline.R,
%   R/stat-timeline-label.R, R/stat-timeline.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\alias{geom_timeline}
\alias{stat_timeline_label}
\alias{stat_timeline}
\title{Geom for NOAA Earthquake dataset}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)

geom_timeline(
  mapping = NULL,
  data = NULL,
  position = "identity",
  show.legend = NA,
  na.rm = FALSE,
  inherit.aes = TRUE,
  ...
)

stat_timeline_label(
  mapping = NULL,
  data = NULL,
  geom = "timeline_label",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)

stat_timeline(
  mapping = NULL,
  data = NULL,
  geom = "timeline",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{geom}{The geometric object to use display the data}
}
\description{
This geom is used to create timelines for the NOAA Significant Earthquakes dataset.
}
\section{Aesthetics}{


\code{geom_timeline()} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}}
  \item \code{min_date}
  \item \code{max_date}
  \item \code{color}
  \item \code{size}
  \item \code{alpha}}
}

\examples{
library(msdrCourseraCapstonePackage)
library(magrittr)
library(ggplot2)
library(dplyr)
library(lubridate)

data_path <- system.file("extdata", "signif.txt", package = "msdrCourseraCapstonePackage")
eq_data_raw <- read.delim(data_path)
eq_data <- eq_clean_data(eq_data_raw)

eq_data_plot <- eq_data \%>\% filter(COUNTRY \%in\% c("CANADA", "USA", "MEXICO"))

# basic timeline of entire dataset
ggplot(eq_data_plot, aes(x = DATE)) + geom_timeline() + theme_eq

# filter timeline with a min date
ggplot(eq_data_plot, aes(x = DATE, min_date = as.Date("1900-01-01"))) + geom_timeline() + theme_eq

# add a max date as well
ggplot(eq_data_plot, aes(x = DATE, min_date = as.Date("1900-01-01"), max_date = as.Date("1950-01-01"))) + geom_timeline() + theme_eq

# use y aesthetic to add COUNTRY stratification
ggplot(eq_data_plot, aes(x = DATE, y = COUNTRY, min_date = as.Date("1900-01-01"), max_date = as.Date("1950-01-01"))) + geom_timeline() + theme_eq

# use size and color of points to represent magnitude and number of casualties, and set alpha to 0.3
plot <- ggplot(eq_data_plot, aes(x = DATE, y = COUNTRY, min_date = as.Date("1900-01-01"), max_date = as.Date("1950-01-01"), size = EQ_PRIMARY, color = TOTAL_DEATHS)) + geom_timeline(alpha = 0.3) + theme_eq
}
