% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msdrCourseraCapstonePackage.R
\name{msdrCourseraCapstone-ggproto}
\alias{msdrCourseraCapstone-ggproto}
\title{Base ggproto classes for msdrCourseraCapstonePackage}
\description{
Base ggproto classes for msdrCourseraCapstonePackage
}
\seealso{
ggproto
}
\keyword{internal}
