% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_eq.R
\docType{data}
\name{theme_eq}
\alias{theme_eq}
\title{Theme to use with geom_timeline}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 67.}
\usage{
theme_eq
}
\description{
Theme to use with geom_timeline
}
\keyword{datasets}
